Update main.py
Just about to fix the "highlight cell" function.  This can then be used to highlight where a person clicks or to highlight all number 3, etc.

A click needs to add a cell to a list of those highlighted.  This is empty most of the time, but contains at least one cell after a click.